requirements.txt
langchain-mcp-adapters
langchain
langchain-openai
python-dotenv
google-workspace-mcp
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
